Adding separate gradientDescent.py
